v1.1.0 1. remove landing/index.vue 2. fix logout
