+++ src/controller/repairapi.js
@@ -93,16 +93,19 @@ export default{
             vm.data.apiCookies = null
             wx.clearStorageSync()
             wx.reLaunch({
-            url: "/pages/login/main"
+                url: "/pages/login/main"
             })
+
         })
 
+        /*
         wx.removeStorage({key: "repairData", success: function (res) {} })
         return new Promise(function (resolve, reject){
             vm.query("logout", "GET").then(r => {
                 resolve(r)
             })
-        })    
+        })   
+        */ 
     },
     getDevices (){
         var vm = this
