+++ src/main.js
@@ -11,7 +11,7 @@ export default {
   // 这个字段走 app.json
   config: {
     // 页面前带有 ^ 符号的，会被编译成首页，其他页面可以选填，我们会自动把 webpack entry 里面的入口页面加进去
-    pages: ['^pages/landing/main', 'pages/login/main',
+    pages: ['^pages/login/main',
             'pages/tickets/new/main', 'pages/tickets/list/main', 'pages/tickets/detail/main',
             'pages/my/info/main'],
     window: {
