--- src/pages/landing/main.js
@@ -1,5 +1,0 @@
-import Vue from 'vue'
-import App from './index'
-
-const app = new Vue(App)
-app.$mount()
