+++ src/pages/tickets/new/main.js
@@ -3,3 +3,8 @@ import App from './index'
 
 const app = new Vue(App)
 app.$mount()
+export default {
+    config: {
+        enablePullDownRefresh: true
+    }
+  }
